Lekka Reorganizacja kodu plus dodanie klas Ad, Cassette i przerzucenie metod playMusic() i casetteChoose() do klasy ChooseCassette
metoda playMusic() i tak jest tymczasowa raczej
