+++ Assets/wojtek/scripts/Ad.cs
@@ -1,0 +1,28 @@
+using UnityEngine;
+
+public class Ad : MonoBehaviour
+{
+    [SerializeField]
+    private string name;
+    [SerializeField]
+    private int timesUsedInDay;
+    [SerializeField]
+    private int gainMoneyHipHop;
+    [SerializeField]
+    private int gainMoneyDisco;
+    [SerializeField]
+    private int gainMoneyRock;
+    [SerializeField]
+    private int gainMoneyMetal;
+
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
