+++ Assets/wojtek/scripts/Ad.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4ef629219f6b17c4bbd2d813eecbea14
