+++ Assets/wojtek/scripts/Cassette.cs
@@ -1,0 +1,31 @@
+using UnityEngine;
+
+public class Cassette : MonoBehaviour
+{
+    [SerializeField]
+    private string name;
+    [SerializeField]
+    private int timesUsedInDay;
+    [SerializeField]
+    private int gainListenersHipHop;
+    [SerializeField]
+    private int gainListenersDisco;
+    [SerializeField]
+    private int gainListenersRock;
+    [SerializeField]
+    private int gainListenersMetal;
+    [SerializeField]
+    AudioClip musicClip;
+
+
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
